KUDU-3230: Fix the issue of hardcode sasl_proto_name
'kudu' is used as sasl service name which is also used as kerberos
principal name when Kerberos is enabled, as a result, users can only
use 'kudu' than other principals. It is not flexible.

It is better allow to customize sasl_krb5_principal_name by specifying other
arguments for kudu-master and kudu-tserver.

Change-Id: Ia1c8f9d9f772d000d9a588e4e9d6028711a62915